refactor: removed the misc export from package and moved the functions directly into example for openUrlAndListenForAuthRedirect
