+++ src/index.ts
@@ -12,7 +12,6 @@ import * as WalletInstanceAttestation from "./wallet-instance-attestation";
 import * as Trust from "./trust";
 import * as WalletInstance from "./wallet-instance";
 import * as Cie from "./cie";
-import * as Misc from "./utils/misc";
 import { AuthorizationDetail, AuthorizationDetails } from "./utils/par";
 import { createCryptoContextFor } from "./utils/crypto";
 import type { IntegrityContext } from "./utils/integrity";
@@ -30,7 +29,6 @@ export {
   AuthorizationDetails,
   fixBase64EncodingOnKey,
   Cie,
-  Misc,
 };
 
 export type { IntegrityContext, AuthorizationContext };
